refactor: move ActionBar - Updated & Render Event to the NpcDialogueModel
